// src/api/models/OrderDetail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {
    Address,
    DeliveryDate,
    OrderItem,
    OrderTracking,
    User,
} from './';

/**
 * 
 * @export
 * @interface OrderDetail
 */
export interface OrderDetail {
    /**
     * 
     * @type {Address}
     * @memberof OrderDetail
     */
    billing?: Address;
    /**
     * 
     * @type {DeliveryDate}
     * @memberof OrderDetail
     */
    deliveryDate?: DeliveryDate;
    /**
     * 
     * @type {string}
     * @memberof OrderDetail
     */
    deliveryType?: OrderDetailDeliveryTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof OrderDetail
     */
    id?: number;
    /**
     * 
     * @type {Date}
     * @memberof OrderDetail
     */
    orderDate?: Date;
    /**
     * 
     * @type {Array<OrderItem>}
     * @memberof OrderDetail
     */
    orderItems?: Array<OrderItem>;
    /**
     * 
     * @type {string}
     * @memberof OrderDetail
     */
    paymentType?: OrderDetailPaymentTypeEnum;
    /**
     * 
     * @type {Address}
     * @memberof OrderDetail
     */
    shipping?: Address;
    /**
     * 
     * @type {number}
     * @memberof OrderDetail
     */
    total?: number;
    /**
     * 
     * @type {OrderTracking}
     * @memberof OrderDetail
     */
    tracking?: OrderTracking;
    /**
     * 
     * @type {User}
     * @memberof OrderDetail
     */
    user?: User;
}

/**
* @export
* @enum {string}
*/
export enum OrderDetailDeliveryTypeEnum {
    DELIVER = 'DELIVER',
    PICKUP = 'PICK_UP',
    MEETUP = 'MEET_UP'
}
/**
* @export
* @enum {string}
*/
export enum OrderDetailPaymentTypeEnum {
    CASH = 'CASH',
    GCASH = 'GCASH'
}